+++ src/lib/models/reflections/declaration.ts
@@ -208,14 +208,6 @@ export class DeclarationReflection extends ContainerReflection {
         );
     }
 
-    /**
-     * Traverse all potential child reflections of this reflection.
-     *
-     * The given callback will be invoked for all children, signatures and type parameters
-     * attached to this reflection.
-     *
-     * @param callback  The callback function that should be applied for each child reflection.
-     */
     override traverse(callback: TraverseCallback) {
         for (const parameter of this.typeParameters?.slice() || []) {
             if (callback(parameter, TraverseProperty.TypeParameter) === false) {