+++ src/lib/models/reflections/parameter.ts
@@ -13,14 +13,6 @@ export class ParameterReflection extends Reflection {
 
     type?: SomeType;
 
-    /**
-     * Traverse all potential child reflections of this reflection.
-     *
-     * The given callback will be invoked for all children, signatures and type parameters
-     * attached to this reflection.
-     *
-     * @param callback  The callback function that should be applied for each child reflection.
-     */
     override traverse(callback: TraverseCallback) {
         if (this.type instanceof ReflectionType) {
             if (
@@ -32,8 +24,6 @@ export class ParameterReflection extends Reflection {
                 return;
             }
         }
-
-        super.traverse(callback);
     }
 
     /**