+++ src/test/converter2/behavior/removeReflection.ts
@@ -3,3 +3,5 @@ export function foo(first: string, second: string, third: string) {
     second;
     third;
 }
+
+export function nested(a: 1 | { a; 1 }) {}